Move chart components to flat structure
